requirements.txt
requests
pandas
pydantic
fastapi
uvicorn
python-dotenv
gunicorn